Adjust titlebar UI and fix prompts
