+++ crates/title_bar/src/platforms.rs
@@ -1,4 +1,3 @@
-pub mod platform_generic;
 pub mod platform_linux;
 pub mod platform_mac;
 pub mod platform_windows;
